fixed CPU load

+++ include/processor.h
@@ -10,7 +10,7 @@ class Processor {
   // TODO: Declare any necessary private members
  private:
   enum Jiff{user, nice, system, idle, iowait, irq, softirq, steal, guest, guest_nice};
-  int _user, _nice, _system, _idle, _iowait, _irq, _softirq, _steal, _guest, _guest_nice;
+  float _user, _nice, _system, _idle, _iowait, _irq, _softirq, _steal, _guest, _guest_nice;
 };
 
 #endif
